# openapi.yaml
openapi: 3.1.0
info:
  version: 0.1.0
  title: Streamingestarr APIs
  description: Internal and admin APIs for Streamingestarr. These power the web interface and admin panel
    and may change at any time.
  contact:
    name: Streamingestarr
  license:
    name: MIT
    url: https://github.com/owncast/owncast/blob/develop/LICENSE
servers:
- url: http://localhost:8080/api
  description: Localhost port 8080
paths:
...
  /chat:
    get:
      summary: Gets a list of chat messages
      operationId: GetChatMessages
      tags:
      - Internal
      - Chat
      parameters:
      - $ref: '#/components/parameters/AccessToken'
      responses:
        '200':
          description: List of chat messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessages'
        '401':
          $ref: '#/components/responses/401'
        '501':
          $ref: '#/components/responses/501'
...
  /chat/messagevisibility:
    post:
      summary: Update chat message visibility
      operationId: UpdateMessageVisibility
      tags:
      - Internal
      - Chat
      parameters:
      - $ref: '#/components/parameters/AccessToken'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageVisibilityUpdate'
      responses:
        '200':
          description: Message visibility updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseAPIResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        default:
          $ref: '#/components/responses/Default'
  /chat/users/setenabled:
    post:
      summary: Enable/disable a user
      operationId: UpdateUserEnabled
      tags:
      - Internal
      - Chat
      parameters:
      - $ref: '#/components/parameters/AccessToken'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                enabled:
                  type: boolean
      responses:
        '200':
          description: User status has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseAPIResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        default:
          $ref: '#/components/responses/Default'
...
  /admin/chat/users/disabled:
    get:
      summary: Get a list of disabled users
      operationId: GetDisabledUsers
      tags:
      - Internal
      - Admin
      - Chat
      security:
      - BasicAuth: []
      responses:
        '200':
          description: List of disabled users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401BasicAuth'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: GetDisabledUsersOptions
      x-internal: true
      tags:
      - Objects
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
  /admin/chat/users/ipbans/create:
    post:
      summary: Ban an IP address
      operationId: BanIPAddress
      tags:
      - Internal
      - Admin
      - Chat
      security:
      - BasicAuth: []
      requestBody:
        $ref: '#/components/requestBodies/AdminConfigValue'
      responses:
        '200':
          description: IP was successfully banned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseAPIResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401BasicAuth'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: BanIPAddressOptions
      x-internal: true
      tags:
      - Objects
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
  /admin/chat/users/ipbans/remove:
    post:
      summary: Remove an IP ban
      operationId: UnbanIPAddress
      tags:
      - Internal
      - Admin
      - Chat
      security:
      - BasicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminConfigValue'
      responses:
        '200':
          description: IP ban was successfully removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseAPIResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401BasicAuth'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: UnbanIPAddressOptions
      x-internal: true
      tags:
      - Objects
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
  /admin/chat/users/ipbans:
    get:
      summary: Get all banned IP addresses
      operationId: GetIPAddressBans
      tags:
      - Internal
      - Admin
      - Chat
      security:
      - BasicAuth: []
      responses:
        '200':
          description: List of banned IP addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IPAddress'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401BasicAuth'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: GetIPAddressBansOptions
      x-internal: true
      tags:
      - Objects
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
  /admin/chat/users/setmoderator:
    post:
      summary: Set moderator status for a user
      operationId: UpdateUserModerator
      tags:
      - Internal
      - Admin
      - Chat
      security:
      - BasicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                isModerator:
                  type: boolean
      responses:
        '200':
          description: Successfully update the moderator status of the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseAPIResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401BasicAuth'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: UpdateUserModeratorOptions
      x-internal: true
      tags:
      - Objects
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
  /admin/chat/users/moderators:
    get:
      summary: Get a list of moderator users
      operationId: GetModerators
      tags:
      - Internal
      - Admin
      - Chat
      security:
      - BasicAuth: []
      responses:
        '200':
          description: List of moderator users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401BasicAuth'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: GetModeratorsOptions
      x-internal: true
      tags:
      - Objects
      - Internal
      - Admin
      responses:
        '204':
          $ref: '#/components/responses/204'
...
  /admin/config/chat/establishedusermode:
    post:
      summary: Enable/disable chat established user mode
      operationId: SetEnableEstablishedChatUserMode
      tags:
      - Internal
      - Admin
      - Chat
      security:
      - BasicAuth: []
      requestBody:
        $ref: '#/components/requestBodies/AdminConfigValue'
      responses:
        '200':
          description: Chat established user mode updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseAPIResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401BasicAuth'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: SetEnableEstablishedChatUserModeOptions
      x-internal: true
      tags:
      - Objects
      - Internal
      - Admin
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
...
  /integrations/chat/system:
    post:
      summary: Send a system message to the chat
      operationId: SendSystemMessage
      tags:
      - External
      - Chat
      security:
      - BearerAuth: []
      requestBody:
        description: The content of the system message
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageEvent'
      responses:
        '200':
          description: Message sent successfully
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseAPIResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: SendSystemMessageOptions
      x-internal: true
      tags:
      - Objects
      - Internal
      - Admin
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
  /integrations/chat/system/client/{clientId}:
    post:
      summary: Send a system message to a single client
      operationId: SendSystemMessageToConnectedClient
      tags:
      - External
      - Chat
      security:
      - BearerAuth: []
      parameters:
      - in: path
        name: clientId
        required: true
        schema:
          type: integer
        description: The ID of the specified client
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemMessage'
      responses:
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: SendSystemMessageToConnectedClientOptions
      x-internal: true
      tags:
      - Objects
      - Internal
      - Admin
      - Chat
      parameters:
      - in: path
        name: clientId
        required: true
        schema:
          type: integer
        description: The ID of the specified client
      responses:
        '204':
          $ref: '#/components/responses/204'
  /integrations/chat/user:
    post:
      deprecated: true
      summary: Send a user message to chat
      operationId: SendUserMessage
      tags:
      - External
      - Chat
      security:
      - BearerAuth: []
      responses:
        '400':
          $ref: '#/components/responses/400'
    options:
      deprecated: true
      operationId: SendUserMessageOptions
      tags:
      - Objects
      - Internal
      - Admin
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
  /integrations/chat/send:
    post:
      summary: Send a message to chat as a specific 3rd party bot/integration based on its access token
      operationId: SendIntegrationChatMessage
      tags:
      - External
      - Chat
      security:
      - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageEvent'
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseAPIResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: SendIntegrationChatMessageOptions
      x-internal: true
      tags:
      - Objects
      - External
      - Admin
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
  /integrations/chat/action:
    post:
      summary: Send a user action to chat
      operationId: SendChatAction
      tags:
      - External
      - Chat
      security:
      - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageEvent'
      responses:
        '200':
          description: Action sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseAPIResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: SendChatActionOptions
      x-internal: true
      tags:
      - Objects
      - Internal
      - Admin
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
  /integrations/chat/messagevisibility:
    post:
      summary: Hide chat message
      operationId: ExternalUpdateMessageVisibility
      tags:
      - External
      - Chat
      security:
      - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageVisibilityUpdate'
      responses:
        '200':
          description: Messages updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseAPIResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401BasicAuth'
        default:
          $ref: '#/components/responses/Default'
    options:
      operationId: ExternalUpdateMessageVisibilityOptions
      x-internal: true
      tags:
      - Objects
      - External
      - Admin
      - Chat
      responses:
        '204':
          $ref: '#/components/responses/204'
...
components:
  schemas:
    Error:
      type: object
      description: Structure for an error response
      properties:
        error:
          type: string
    BaseAPIResponse:
      type: object
      description: Simple API response
      properties:
        message:
          type: string
        success:
          type: boolean
...
    User:
      type: object
      properties:
        createdAt:
          type: string
        disabledAt:
          type: string
        nameChangedAt:
          type: string
        id:
          type: string
        displayName:
          type: string
        previousNames:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        displayColor:
          type: integer
        isBot:
          type: boolean
        authenticated:
          type: boolean
    Users:
      type: array
      items:
        $ref: '#/components/schemas/User'
    Event:
      type: object
      properties:
        timestamp:
          type: string
        type:
          type: string
        id:
          type: string
    UserEvent:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        hiddenAt:
          type: string
        clientId:
          type: integer
    MessageEvent:
      type: object
      properties:
        body:
          type: string
    ChatMessages:
      type: array
      items:
        anyOf:
        - $ref: '#/components/schemas/UserMessage'
        - $ref: '#/components/schemas/SystemMessage'
        - $ref: '#/components/schemas/ActionMessage'
    UserMessage:
      type: object
      allOf:
      - $ref: '#/components/schemas/Event'
      - $ref: '#/components/schemas/UserEvent'
      - $ref: '#/components/schemas/MessageEvent'
    SystemMessage:
      type: object
      allOf:
      - $ref: '#/components/schemas/Event'
      - $ref: '#/components/schemas/MessageEvent'
...
    ActionMessage:
      type: object
      allOf:
      - $ref: '#/components/schemas/Event'
      - $ref: '#/components/schemas/MessageEvent'
...
    IPAddress:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        ipAddress:
          type: string
        notes:
          type: string
...
    MessageVisibilityUpdate:
      type: object
      properties:
        idArray:
          type: array
          items:
            type: string
        visible:
          type: boolean
...
    AdminConfigValue:
      type: object
      properties:
        value:
          oneOf:
          - type: string
          - type: boolean
          - type: array
            items:
              type: string
          - type: number
  headers:
    Access-Control-Allow-Origin:
      schema:
        type: string
    Access-Control-Allow-Credentials:
      schema:
        type: string
        format: boolean
    Access-Control-Allow-Headers:
      schema:
        type: string
    WWW-Authenticate:
      schema:
        type: string
    X-Content-Type-Options:
      schema:
        type: string
  requestBodies:
    AdminConfigValue:
      description: A single value that is used in the request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminConfigValue'
...
  responses:
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseAPIResponse'
    '204':
      description: No Content
...
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseAPIResponse'
    '401':
      description: Unauthorized
      content:
        text/plain:
          schema:
            type: string
    401BasicAuth:
      description: Basic auth failure
      headers:
        Access-Control-Allow-Origin:
          $ref: '#/components/headers/Access-Control-Allow-Origin'
        Access-Control-Allow-Credentials:
          $ref: '#/components/headers/Access-Control-Allow-Credentials'
        Access-Control-Allow-Headers:
          $ref: '#/components/headers/Access-Control-Allow-Headers'
        WWW-Authenticate:
          $ref: '#/components/headers/WWW-Authenticate'
        X-Content-Type-Options:
          $ref: '#/components/headers/X-Content-Type-Options'
      content:
        text/plain:
          schema:
            type: string
...
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    '501':
      description: Not Implemented
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
...
    AccessToken:
      in: query
      name: accessToken
      schema:
        type: string
      required: true
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
...